API: added HomologyAggregator.commit() method
[CHANGED] commits the cached indices into their respective tables
